Fancy: Hoist statics for styled components
This update bumps Fancy to the latest v2.0.5, which hoists non React
statics to the wrapped Styled component.

This is expected behaviour for HOC wrappers, but is missing from (stock)
Emotion.

Certain Components were refactored to better leverage the static hoists.
